refatoracao de funcionalidades e reestruturacao de pastas e arquivos
