+++ index.html
@@ -5,11 +5,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quiz</title>
-    <script src="./scripts/modo-escuro.js" defer></script>
     <script src="./scripts/script.js" defer type="module"></script>
-    <script src="./scripts/cronometro.js" defer type="module"></script>
-    <script src="./scripts/calculaMedia.js" defer type="module"></script>
-    <script src="./scripts/ranking.js" defer type="module"></script>
     <link rel="stylesheet" href="css/main.css" />
     <link rel="stylesheet"
         href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
@@ -20,7 +16,7 @@
 
     <section class="audio">
         <audio id="meuAudio">
-            <source src="./audio-arquivo/sound.mp3" type="audio/mpeg">
+            <source src="./assets/audio/sound.mp3" type="audio/mpeg">
             Seu navegador não possui suporte para audio, tente outro
             navegador.
         </audio>
@@ -82,6 +78,7 @@
 
                 </tbody>
             </table>
+            
             <h1>Média de Acertos</h1>
             <table id="tabela-medias-acertos">
                 <thead>
@@ -138,7 +135,7 @@
                 </thead>
                 <tbody></tbody>
             </table>
-    </section>
+        </section>
 
 </body>
