+++ scripts/module/infoUsuario.js
@@ -1,0 +1,37 @@
+export let infoUsuarios = [
+    {
+        nome: 'Ana',
+        temaSelecionado: 'livros',
+        acertos: 8, erros: 2,
+        totalSegundos: 187,
+        dataHoraPreenchimento: '25/07/2023 12:32'
+    },
+    {
+        nome: 'João',
+        temaSelecionado: 'livros',
+        acertos: 5, erros: 5,
+        totalSegundos: 233,
+        dataHoraPreenchimento: '01/07/2023 17:49'
+    },
+    {
+        nome: 'Gabriel',
+        temaSelecionado: 'filmes',
+        acertos: 3, erros: 7,
+        totalSegundos: 267,
+        dataHoraPreenchimento: '10/07/2023 10:15'
+    },
+    {
+        nome: 'Beatriz',
+        temaSelecionado: 'filmes',
+        acertos: 7, erros: 3,
+        totalSegundos: 199,
+        dataHoraPreenchimento: '22/07/2023 21:30'
+    },
+    {
+        nome: 'Maria',
+        temaSelecionado: 'series',
+        acertos: 10, erros: 0,
+        totalSegundos: 212,
+        dataHoraPreenchimento: '29/07/2023 08:45'
+    }
+];
