+++ scripts/module/resultado/calculaMedia.js
